+++ Sources/Variable.swift
@@ -78,6 +78,9 @@ public struct Variable : Equatable, Resolvable {
           current = dictionary.count
         } else {
           current = dictionary[bit]
+          if current == nil, let key = (try? Variable(bit).resolve(context)) as? String {
+            current = dictionary[key]
+          }
         }
       } else if let array = current as? [Any] {
         if let index = Int(bit) {
@@ -92,6 +95,12 @@ public struct Variable : Equatable, Resolvable {
           current = array.last
         } else if bit == "count" {
           current = array.count
+        } else if let index = (try? Variable(bit).resolve(context)) as? Int {
+          if index >= 0 && index < array.count {
+            current = array[index]
+          } else {
+            current = nil
+          }
         }
       } else if let object = current as? NSObject {  // NSKeyValueCoding
 #if os(Linux)
@@ -100,7 +109,11 @@ public struct Variable : Equatable, Resolvable {
         current = object.value(forKey: bit)
 #endif
       } else if let value = current {
-        current = Mirror(reflecting: value).getValue(for: bit)
+        let mirror = Mirror(reflecting: value)
+        current = mirror.getValue(for: bit)
+        if current == nil, let label = (try? Variable(bit).resolve(context)) as? String {
+          current = mirror.getValue(for: label)
+        }
         if current == nil {
           return nil
         }
@@ -176,7 +189,7 @@ func parseFilterComponents(token: String) -> (String, [Variable]) {
 
 extension Mirror {
   func getValue(for key: String) -> Any? {
-    let result = descendant(key)
+    let result = descendant(key) ?? Int(key).flatMap({ descendant($0) })
     if result == nil {
       // go through inheritance chain to reach superclass properties
       return superclassMirror?.getValue(for: key)